Move graph.rb test flow
From test/fixtures/graph_presenter_test to test/fixtures/smart_answer_flows.

I'm moving all Smart Answer test flows and locale files to
test/fixtures/smart_answer_flows. The majority of test flows and locale files
already live in that directory but there are also a small number of them living
in test/fixtures or test/fixtures/a-n-other-directory.
