+++ test/unit/graph_presenter_test.rb
@@ -1,14 +1,14 @@
 require_relative '../test_helper'
 require_relative '../helpers/i18n_test_helper'
 
-require 'fixtures/graph_presenter_test/graph'
+require 'fixtures/smart_answer_flows/graph'
 
 module SmartAnswer
   class GraphPresenterTest < ActiveSupport::TestCase
     include I18nTestHelper
 
     setup do
-      use_additional_translation_file(fixture_file('graph_presenter_test/graph.yml'))
+      use_additional_translation_file(fixture_file('smart_answer_flows/locales/en/graph.yml'))
       @flow = SmartAnswer::GraphFlow.build
       @presenter = GraphPresenter.new(@flow)
     end
